+++ packages/wrangler/src/api/startDevWorker/ConfigController.ts
@@ -401,6 +401,7 @@ async function resolveConfig(
 		projectRoot: entry.projectRoot,
 		bindings,
 		migrations: input.migrations ?? config.migrations,
+		exports: input.exports ?? config.exports,
 		sendMetrics: input.sendMetrics ?? config.send_metrics,
 		triggers: await resolveTriggers(config, input),
 		env: input.env,
@@ -519,7 +520,8 @@ async function resolveConfig(
 
 		// TODO(do) support remote wrangler dev
 		const classNameToUseSQLite = getDurableObjectClassNameToUseSQLiteMap(
-			resolved.migrations
+			resolved.migrations,
+			resolved.exports
 		);
 		if (
 			resolved.dev.remote &&